requirements.txt
streamlit>=1.20
flask>=2.0
pillow>=9.0
numpy>=1.24
opencv-python-headless>=4.7
pandas>=1.5
PyMuPDF>=1.22
tqdm>=4.64
